Renamed from _export to _exporter
Some styling
