+++ datacontract/export/exporter_factory.py
@@ -91,7 +91,7 @@ exporter_factory.register_lazy_exporter(
 
 exporter_factory.register_lazy_exporter(
     name=ExportFormat.mermaid,
-    module_path="datacontract.export.mermaid_export",
+    module_path="datacontract.export.mermaid_exporter",
     class_name="MermaidExporter",
 )
 
@@ -133,7 +133,7 @@ exporter_factory.register_lazy_exporter(
 
 exporter_factory.register_lazy_exporter(
     name=ExportFormat.html,
-    module_path="datacontract.export.html_export",
+    module_path="datacontract.export.html_exporter",
     class_name="HtmlExporter",
 )
