+++ datacontract/templates/datacontract.html
@@ -85,7 +85,7 @@
             </div>
             <div class="mt-3">
               <div class="overflow-hidden bg-white shadow-sm ring-1 ring-gray-900/5 sm:rounded-lg">
-                <div class="diagram-container p-4" id="diagram-container" style="width: 100%; height: 520px;">
+                <div class="diagram-container p-4 w-full" id="diagram-container">
                   <pre class="mermaid">
                     {{ mermaid_diagram }}
                   </pre>
